refactor(web): route format/file-watch through EditorService + rename cli/hooks/cli-skills
F. The workspace router was reaching into services/formatter and
   services/file-watcher directly. Both already had matching methods on
   EditorService (formatFile, subscribeToFileChanges); switched the
   router to call them via editorService so the public seam for the
   "editor" orchestration matches the doc table entry.

G. Rename services/{cli,hooks,cli-skills}.ts to {cli,hooks,cli-skills}-
   service.ts to match the doc's `{domain}-service.ts` naming rule.
   Importers in cli/hooks routers and setup-service updated in lock-step.

Refs: https://github.com/band-app/band/issues/535
